Add Client method with postcode suggestion implemented. Necessery classes added (AddressDao, ClientService etc.)

+++ src/main/java/pl/sda/main/Main.java
@@ -1,30 +1,28 @@
 package pl.sda.main;
-
 import org.hibernate.query.Query;
-import pl.sda.dao.EmployeeDao;
 import pl.sda.entities.*;
 import org.hibernate.Session;
 import pl.sda.enums.Category;
+import pl.sda.service.AddressService;
+import pl.sda.service.ClientService;
 import pl.sda.service.EmployeeService;
 import pl.sda.util.HibernateUtil;
-
 import java.math.BigDecimal;
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-
 import static java.lang.System.in;
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import org.hibernate.cfg.Configuration;
+
 
 public class Main {
     private static Scanner scanner = new Scanner(in);
     private static EmployeeService employeeService = new EmployeeService();
-    private  static EmployeeDao employeeDao = new EmployeeDao();
+    private static ClientService clientService = new ClientService();
+    private static AddressService addressService = new AddressService();
 
     public static void main(String[] args) {
+
+        // ********************************DEVELOPMENT TESTS
         Session session = HibernateUtil.getSession();
         Employee employee1 = new Employee("Joanna", "Mickiewicz", "96121536981");
         Client client = new Client("Adam", "Spadam", "59874532145");
@@ -42,6 +40,8 @@ public class Main {
         session.persist(good);
         session.getTransaction().commit();
         session.close();
+
+//        addressService.checkPostcode("Ełk");
         //*******************************************************
 
 
@@ -50,18 +50,15 @@ public class Main {
         do {
             System.out.println("Wybierz opcję: ");
             System.out.println("0 - zakończ");
-            System.out.println("DOT. PRACOWNIKA");
-            System.out.println("11 - Dodaj pracownika");
-            System.out.println("12 - Zmień dane pracownika"); //TODO
-            System.out.println("13 - Usuń pracownika");
-//            System.out.println("DOT. KLIENTA");
-//            System.out.println("21 - Dodaj klienta");
-//            System.out.println("22 - Zmień dane klienta"); //TODO
-//            System.out.println("23 - Usuń klienta");
+//            System.out.println("DOT. PRACOWNIKA");
+//            System.out.println("11 - Dodaj pracownika");
+//            System.out.println("12 - Zmień dane pracownika");
+//            System.out.println("13 - Usuń pracownika");
+            System.out.println("DOT. KLIENTA");
+            System.out.println("21 - Dodaj klienta"); //TODO sprawdzić wyświetlanie adresów przypisanych do klienta
+            System.out.println("22 - Zmień dane klienta");
+            System.out.println("23 - Usuń klienta");
 //            System.out.println("DOT. UMOWY");
-//            System.out.println("21 - Dodaj umowę");
-//            System.out.println("22 - anuluj umowę");
-//            System.out.println("DOT. TOWARU");
 //            System.out.println("21 - Dodaj umowę");
 //            System.out.println("22 - anuluj umowę");
 //            System.out.println("DOT. WYSZUKIWANIA");
@@ -75,14 +72,56 @@ public class Main {
             switch (choose) {
                 case 0: HibernateUtil.closeSessionFactory(); System.exit(0);
                 case 11: addEmployee(); break;
-                case 12: editEmployee(); break; //TODO
+                case 12: editEmployee(); break; 
                 case 13: deleteEmployeeByPesel(); break;
-//                case 21: addClient(); break;
+                case 21: addClient(); break;
 //                case 23: deleteClient(); break;
 //                case 31: findContractById(); break;
                 default: System.out.println("Nierozpoznana operacja"); break;
             }
         } while (true);
+    }
+
+    private static void addClient() {
+        System.out.println("Podaj imię: ");
+        String name = scanner.next();
+
+        System.out.println("Podaj nazwisko: ");
+        String surname = scanner.next();
+
+        System.out.println("Podaj PESEL: ");
+        String pesel = scanner.next();
+
+        System.out.println("Podaj miasto: ");
+        String city = scanner.next();
+
+        System.out.println("Podaj ulicę: ");
+        String street = scanner.next();
+
+        System.out.println("Podaj numer domu: ");
+        String houseNum = scanner.next();
+
+        System.out.println("Podaj numer mieszkania lub wpisz --- by pominąć: ");
+        String flatNum = scanner.next();
+
+        System.out.println(addressService.checkPostcode(city));
+
+        System.out.println("Podaj numer telefonu lub wpisz --- by pominąć");
+        String phone = scanner.next();
+
+        System.out.println("Podaj email lub wpisz --- by pominąć");
+        String email = scanner.next();
+
+        System.out.println("Podaj dodatkowe uwagi lub wpisz --- by pominąć");
+        String comments = scanner.next();
+
+        boolean result = clientService.createClientWithAddress(name, surname, pesel, city, street, houseNum, flatNum, phone, email, comments);
+
+        if (result) {
+            System.out.println("Pomyślnie utworzono klienta: " + name + " " + surname);
+        } else {
+            System.out.println("Nie utworzono klienta: "  + name + " " + surname);
+        }
     }
 
     private static void editEmployee() {
@@ -113,38 +152,27 @@ public class Main {
         query2.executeUpdate();
         session.getTransaction().commit();
         session.close();
-
-
-
-//        Query query1 = session.createQuery("FROM Employee WHERE pesel = '" + newPesel + "' ");
-//        Query query2 = session.createQuery("UPDATE Employee e SET e.name = '" + newName + "' "+
-//                "WHERE pesel = '" + newPesel + "' ");
-//
-//        query2.executeUpdate();
-
-
-
-        System.out.println("Nie zaktualizowano bazy ;/ na później");
-        session.close();
-
     }
 
     private static void deleteEmployeeByPesel() {
 
         System.out.println("Podaj PESEL usuwanego pracownika");
         String pesel = scanner.next();
-
-        Session session = HibernateUtil.getSession();
-        Query delete = session.createQuery("DELETE FROM Employee p where p.pesel=:pesel")
-                .setParameter("pesel", pesel);
-        session.getTransaction().begin();
-        delete.executeUpdate();
-        session.getTransaction().commit();
-        session.close();
-        System.out.println("Pomyślnie usunięto pracownika o PESEL: " + pesel);
-    }
-
-
+        if (employeeService.existsByPesel(pesel)){
+            Session session = HibernateUtil.getSession();
+            Query delete = session.createQuery("DELETE FROM Employee p where p.pesel=:pesel")
+                    .setParameter("pesel", pesel);
+            session.getTransaction().begin();
+            delete.executeUpdate();
+            session.getTransaction().commit();
+            session.close();
+            System.out.println("Pomyślnie usunięto pracownika o PESEL: " + pesel);
+        } else {
+            System.out.println("Nie odnaleziono pracownika o podanym numerze PESEL");
+        }
+     }
+     
+     
     private static void addEmployee() {
         System.out.println("Podaj imię: ");
         String name = scanner.next();
@@ -154,7 +182,11 @@ public class Main {
 
         System.out.println("Podaj PESEL: ");
         String pesel = scanner.next();
-
+        
+        while(employeeService.existsByPesel(pesel)) {
+            System.out.println("Podany PESEL jest przypisany do istniejącego pracownika. Podaj nowy PESEL: ");
+            pesel = scanner.next();
+        }
 
         boolean result = employeeService.createEmployeeByPesel(name, surname, pesel);
